requirements.txt
awscli-local==0.20
boto3==1.24.96
botocore==1.27.96
cachetools==5.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
dill==0.3.2
dnslib==0.9.23
dnspython==2.3.0
ecdsa==0.18.0
idna==3.4
jmespath==1.0.1
localstack==1.3.1
localstack-client==1.39
localstack-ext==1.3.1
markdown-it-py==2.1.0
mdurl==0.1.2
packaging==23.0
pbr==5.11.1
plux==1.3.1
psutil==5.9.4
psycopg2-binary==2.9.5
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
PyYAML==6.0
requests==2.28.2
rich==13.3.1
rsa==4.9
s3transfer==0.6.0
semver==2.13.0
six==1.16.0
stevedore==4.1.1
tabulate==0.9.0
tailer==0.4.1
tomli==2.0.1
urllib3==1.26.14
